refactor get resource links
